requirements.txt
django
djangorestframework
markdown
django-filter
django-environ
psycopg2-binary